requirements.txt
beautifulsoup4==4.12.3
selenium==4.22.0
python-telegram-bot==21.3
undetected-chromedriver==3.5.3
python-dotenv==1.0.1